requirements.txt
scrapy==1.2.1
tablib
googlemaps==2.4.4
pyexcel-io
pyexcel
pymongo
geopy==1.11.0
python-Levenshtein